requirements.txt
mysql-connector-python==8.0.22
PyMySQL==1.0.2
SQLAlchemy==1.3.12
